+++ ArtemisKit/Sources/Dashboard/DashboardViewModel.swift
@@ -2,20 +2,14 @@ import Common
 import Foundation
 import SharedModels
 import SharedServices
-import UserStore
 
-@MainActor
-class DashboardViewModel: ObservableObject {
+class DashboardViewModel: BaseViewModel {
 
     @Published var coursesForDashboard: DataState<[CourseForDashboard]> = DataState.loading
-    @Published var error: UserFacingError? {
-        didSet {
-            showError = error != nil
-        }
-    }
-    @Published var showError = false
 
-    init() {
+    override init() {
+        super.init()
+
         Task {
             await loadCourses()
         }
